\name{mancie_example}
\docType{data}
\alias{mancie_example}
\alias{exp}
\alias{ann_exp}
\alias{DNase}
\alias{ann_DNase}

\title{ The demo dataset for the \code{MANCIE} package
}
\description{ This demo dataset is a small portion of the Encode dataset used in our publication.
}
\usage{data(mancie_example)}
\format{  4 data frames
}

\keyword{datasets}
